backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
postgrest>=2.0.0,<3.0.0
msal>=1.24.0
requests>=2.31.0